problem: messy cli app configuration vars and setters
solution: refactor to own files and build some simple fns to
handle flags and cmds assignments
