+++ cmd/geth/main.go
@@ -19,7 +19,6 @@ package main
 
 import (
 	"fmt"
-	"log"
 	"math/rand"
 	"os"
 	"path/filepath"
@@ -30,10 +29,7 @@ import (
 	"github.com/ethereumproject/benchmark/rtprof"
 	"github.com/ethereumproject/go-ethereum/common"
 	"github.com/ethereumproject/go-ethereum/console"
-	"github.com/ethereumproject/go-ethereum/core"
-	"github.com/ethereumproject/go-ethereum/eth"
 	"github.com/ethereumproject/go-ethereum/logger"
-	"github.com/ethereumproject/go-ethereum/metrics"
 )
 
 // Version is the application revision identifier. It can be set with the linker
@@ -45,67 +41,6 @@ func init() {
 	common.SetClientVersion(Version)
 }
 
-var makeDagCommand = cli.Command{
-	Action:  makedag,
-	Name:    "make-dag",
-	Aliases: []string{"makedag"},
-	Usage:   "Generate ethash dag (for testing)",
-	Description: `
-		The makedag command generates an ethash DAG in /tmp/dag.
-
-		This command exists to support the system testing project.
-		Regular users do not need to execute it.
-				`,
-}
-
-var gpuInfoCommand = cli.Command{
-	Action:  gpuinfo,
-	Name:    "gpu-info",
-	Aliases: []string{"gpuinfo"},
-	Usage:   "GPU info",
-	Description: `
-	Prints OpenCL device info for all found GPUs.
-			`,
-}
-
-var gpuBenchCommand = cli.Command{
-	Action:  gpubench,
-	Name:    "gpu-bench",
-	Aliases: []string{"gpubench"},
-	Usage:   "Benchmark GPU",
-	Description: `
-	Runs quick benchmark on first GPU found.
-			`,
-}
-
-var versionCommand = cli.Command{
-	Action: version,
-	Name:   "version",
-	Usage:  "Print ethereum version numbers",
-	Description: `
-	The output of this command is supposed to be machine-readable.
-			`,
-}
-
-var makeMlogDocCommand = cli.Command{
-	Action: makeMLogDocumentation,
-	Name:   "mdoc",
-	Usage:  "Generate mlog documentation",
-	Description: `
-	Auto-generates documentation for all available mlog lines.
-	Use -md switch to toggle markdown output (eg. for wiki).
-	Arguments may be used to specify exclusive candidate components;
-	so 'geth mdoc -md discover' will generate markdown documentation only
-	for the 'discover' component.
-			`,
-	Flags: []cli.Flag{
-		cli.BoolFlag{
-			Name:  "md",
-			Usage: "Toggle markdown formatting",
-		},
-	},
-}
-
 func makeCLIApp() (app *cli.App) {
 	app = cli.NewApp()
 	app.Name = filepath.Base(os.Args[0])
@@ -114,208 +49,10 @@ func makeCLIApp() (app *cli.App) {
 	app.Action = geth
 	app.HideVersion = true // we have a command to print the version
 
-	app.Commands = []cli.Command{
-		importCommand,
-		exportCommand,
-		dumpChainConfigCommand,
-		upgradedbCommand,
-		dumpCommand,
-		rollbackCommand,
-		recoverCommand,
-		resetCommand,
-		monitorCommand,
-		accountCommand,
-		walletCommand,
-		consoleCommand,
-		attachCommand,
-		javascriptCommand,
-		statusCommand,
-		apiCommand,
-		makeDagCommand,
-		gpuInfoCommand,
-		gpuBenchCommand,
-		versionCommand,
-		makeMlogDocCommand,
-		buildAddrTxIndexCommand,
-	}
+	setAppCommands(app, appCmds)
+	setAppFlagsByGroup(app, AppHelpFlagAndCommandGroups)
 
-	app.Flags = []cli.Flag{
-		PprofFlag,
-		PprofIntervalFlag,
-		SputnikVMFlag,
-		NodeNameFlag,
-		UnlockedAccountFlag,
-		PasswordFileFlag,
-		AccountsIndexFlag,
-		BootnodesFlag,
-		DataDirFlag,
-		DocRootFlag,
-		KeyStoreDirFlag,
-		ChainIdentityFlag,
-		BlockchainVersionFlag,
-		FastSyncFlag,
-		AddrTxIndexFlag,
-		AddrTxIndexAutoBuildFlag,
-		CacheFlag,
-		LightKDFFlag,
-		JSpathFlag,
-		ListenPortFlag,
-		MaxPeersFlag,
-		MaxPendingPeersFlag,
-		EtherbaseFlag,
-		GasPriceFlag,
-		MinerThreadsFlag,
-		MiningEnabledFlag,
-		MiningGPUFlag,
-		AutoDAGFlag,
-		TargetGasLimitFlag,
-		NATFlag,
-		NatspecEnabledFlag,
-		NoDiscoverFlag,
-		NodeKeyFileFlag,
-		NodeKeyHexFlag,
-		RPCEnabledFlag,
-		RPCListenAddrFlag,
-		RPCPortFlag,
-		RPCApiFlag,
-		WSEnabledFlag,
-		WSListenAddrFlag,
-		WSPortFlag,
-		WSApiFlag,
-		WSAllowedOriginsFlag,
-		IPCDisabledFlag,
-		IPCApiFlag,
-		IPCPathFlag,
-		ExecFlag,
-		PreloadJSFlag,
-		WhisperEnabledFlag,
-		DevModeFlag,
-		TestNetFlag,
-		NetworkIdFlag,
-		RPCCORSDomainFlag,
-		NeckbeardFlag,
-		VerbosityFlag,
-		DisplayFlag,
-		DisplayFormatFlag,
-		VModuleFlag,
-		LogDirFlag,
-		LogMaxSizeFlag,
-		LogMinSizeFlag,
-		LogMaxTotalSizeFlag,
-		LogIntervalFlag,
-		LogMaxAgeFlag,
-		LogCompressFlag,
-		LogStatusFlag,
-		MLogFlag,
-		MLogDirFlag,
-		MLogComponentsFlag,
-		BacktraceAtFlag,
-		MetricsFlag,
-		FakePoWFlag,
-		SolcPathFlag,
-		GpoMinGasPriceFlag,
-		GpoMaxGasPriceFlag,
-		GpoFullBlockRatioFlag,
-		GpobaseStepDownFlag,
-		GpobaseStepUpFlag,
-		GpobaseCorrectionFactorFlag,
-		ExtraDataFlag,
-		Unused1,
-	}
-
-	app.Before = func(ctx *cli.Context) error {
-
-		// It's a patch.
-		// Don't know why urfave/cli isn't catching the unknown command on its own.
-		if ctx.Args().Present() {
-			commandExists := false
-			for _, cmd := range app.Commands {
-				if cmd.HasName(ctx.Args().First()) {
-					commandExists = true
-				}
-			}
-			if !commandExists {
-				if e := cli.ShowCommandHelp(ctx, ctx.Args().First()); e != nil {
-					return e
-				}
-			}
-		}
-
-		// Check for --exec set without console OR attach
-		if ctx.IsSet(ExecFlag.Name) {
-			// If no command is used, OR command is not one of the valid commands attach/console
-			if cmdName := ctx.Args().First(); cmdName == "" || (cmdName != "console" && cmdName != "attach") {
-				log.Printf("Error: --%v flag requires use of 'attach' OR 'console' command, command was: '%v'", ExecFlag.Name, cmdName)
-				cli.ShowCommandHelp(ctx, consoleCommand.Name)
-				cli.ShowCommandHelp(ctx, attachCommand.Name)
-				os.Exit(1)
-			}
-		}
-
-		if ctx.IsSet(SputnikVMFlag.Name) {
-			if core.SputnikVMExists {
-				core.UseSputnikVM = true
-			} else {
-				log.Fatal("This version of geth wasn't built to include SputnikVM. To build with SputnikVM, use -tags=sputnikvm following the go build command.")
-			}
-		}
-
-		// Check for migrations and handle if conditionals are met.
-		if err := handleIfDataDirSchemaMigrations(ctx); err != nil {
-			return err
-		}
-
-		if err := setupLogRotation(ctx); err != nil {
-			return err
-		}
-
-		// Handle parsing and applying log verbosity, severities, and default configurations from context.
-		if err := setupLogging(ctx); err != nil {
-			return err
-		}
-
-		// Handle parsing and applying log rotation configs from context.
-		if err := setupLogRotation(ctx); err != nil {
-			return err
-		}
-
-		if s := ctx.String("metrics"); s != "" {
-			go metrics.CollectToFile(s)
-		}
-
-		// This should be the only place where reporting is enabled
-		// because it is not intended to run while testing.
-		// In addition to this check, bad block reports are sent only
-		// for chains with the main network genesis block and network id 1.
-		eth.EnableBadBlockReporting = true
-
-		// (whilei): I use `log` instead of `glog` because git diff tells me:
-		// > The output of this command is supposed to be machine-readable.
-		gasLimit := ctx.GlobalString(aliasableName(TargetGasLimitFlag.Name, ctx))
-		if _, ok := core.TargetGasLimit.SetString(gasLimit, 0); !ok {
-			return fmt.Errorf("malformed %s flag value %q", aliasableName(TargetGasLimitFlag.Name, ctx), gasLimit)
-		}
-
-		// Set morden chain by default for dev mode.
-		if ctx.GlobalBool(aliasableName(DevModeFlag.Name, ctx)) {
-			if !ctx.GlobalIsSet(aliasableName(ChainIdentityFlag.Name, ctx)) {
-				if e := ctx.Set(aliasableName(ChainIdentityFlag.Name, ctx), "morden"); e != nil {
-					return fmt.Errorf("failed to set chain value: %v", e)
-				}
-			}
-		}
-
-		if port := ctx.GlobalInt(PprofFlag.Name); port != 0 {
-			interval := 5 * time.Second
-			if i := ctx.GlobalInt(PprofIntervalFlag.Name); i > 0 {
-				interval = time.Duration(i) * time.Second
-			}
-			rtppf.Start(interval, port)
-		}
-
-		return nil
-	}
-
+	app.Before = appBeforeContext
 	app.After = func(ctx *cli.Context) error {
 		rtppf.Stop()
 		logger.Flush()
