+++ cmd/geth/usage.go
@@ -50,182 +50,6 @@ COPYRIGHT:
    {{end}}
 `
 
-// flagGroup is a collection of commands and/or flags belonging to a single topic.
-type flagGroup struct {
-	Name     string
-	Flags    []cli.Flag
-	Commands []cli.Command
-}
-
-// AppHelpFlagAndCommandGroups is the application flags, grouped by functionality.
-var AppHelpFlagAndCommandGroups = []flagGroup{
-	{
-		Name: "ETHEREUM",
-		Commands: []cli.Command{
-			importCommand,
-			exportCommand,
-			dumpChainConfigCommand,
-			dumpCommand,
-			rollbackCommand,
-			recoverCommand,
-			resetCommand,
-		},
-		Flags: []cli.Flag{
-			DataDirFlag,
-			ChainIdentityFlag,
-			NetworkIdFlag,
-			DevModeFlag,
-			NodeNameFlag,
-			FastSyncFlag,
-			CacheFlag,
-			LightKDFFlag,
-			SputnikVMFlag,
-			BlockchainVersionFlag,
-		},
-	},
-	{
-		Name: "ACCOUNT",
-		Commands: []cli.Command{
-			accountCommand,
-			walletCommand,
-			buildAddrTxIndexCommand,
-		},
-		Flags: []cli.Flag{
-			KeyStoreDirFlag,
-			UnlockedAccountFlag,
-			PasswordFileFlag,
-			AccountsIndexFlag,
-			AddrTxIndexFlag,
-			AddrTxIndexAutoBuildFlag,
-		},
-	},
-	{
-		Name: "API AND CONSOLE",
-		Commands: []cli.Command{
-			consoleCommand,
-			attachCommand,
-			javascriptCommand,
-			apiCommand,
-		},
-		Flags: []cli.Flag{
-			RPCEnabledFlag,
-			RPCListenAddrFlag,
-			RPCPortFlag,
-			RPCApiFlag,
-			WSEnabledFlag,
-			WSListenAddrFlag,
-			WSPortFlag,
-			WSApiFlag,
-			WSAllowedOriginsFlag,
-			IPCDisabledFlag,
-			IPCApiFlag,
-			IPCPathFlag,
-			RPCCORSDomainFlag,
-			JSpathFlag,
-			ExecFlag,
-			PreloadJSFlag,
-		},
-	},
-	{
-		Name: "NETWORKING",
-		Flags: []cli.Flag{
-			BootnodesFlag,
-			ListenPortFlag,
-			MaxPeersFlag,
-			MaxPendingPeersFlag,
-			NATFlag,
-			NoDiscoverFlag,
-			NodeKeyFileFlag,
-			NodeKeyHexFlag,
-		},
-	},
-	{
-		Name: "MINER",
-		Commands: []cli.Command{
-			makeDagCommand,
-		},
-		Flags: []cli.Flag{
-			MiningEnabledFlag,
-			MinerThreadsFlag,
-			MiningGPUFlag,
-			AutoDAGFlag,
-			EtherbaseFlag,
-			TargetGasLimitFlag,
-			GasPriceFlag,
-			ExtraDataFlag,
-		},
-	},
-	{
-		Name: "GAS PRICE ORACLE",
-		Flags: []cli.Flag{
-			GpoMinGasPriceFlag,
-			GpoMaxGasPriceFlag,
-			GpoFullBlockRatioFlag,
-			GpobaseStepDownFlag,
-			GpobaseStepUpFlag,
-			GpobaseCorrectionFactorFlag,
-		},
-	},
-	{
-		Name: "LOGGING AND DEBUGGING",
-		Commands: []cli.Command{
-			versionCommand,
-			statusCommand,
-			monitorCommand,
-			makeMlogDocCommand,
-			gpuInfoCommand,
-			gpuBenchCommand,
-		},
-		Flags: []cli.Flag{
-			VerbosityFlag,
-			VModuleFlag,
-			LogDirFlag,
-			LogMaxSizeFlag,
-			LogMinSizeFlag,
-			LogMaxTotalSizeFlag,
-			LogIntervalFlag,
-			LogMaxAgeFlag,
-			LogCompressFlag,
-			LogStatusFlag,
-			DisplayFlag,
-			DisplayFormatFlag,
-			NeckbeardFlag,
-			MLogFlag,
-			MLogDirFlag,
-			MLogComponentsFlag,
-			BacktraceAtFlag,
-			MetricsFlag,
-			FakePoWFlag,
-		},
-	},
-	{
-		Name: "EXPERIMENTAL",
-		Flags: []cli.Flag{
-			WhisperEnabledFlag,
-			NatspecEnabledFlag,
-			DisplayFlag,
-			DisplayFormatFlag,
-			NeckbeardFlag,
-		},
-	},
-	{
-		Name: "LEGACY",
-		Commands: []cli.Command{
-			upgradedbCommand,
-		},
-		Flags: []cli.Flag{
-			TestNetFlag,
-			Unused1,
-		},
-	},
-	{
-		Name: "MISCELLANEOUS",
-		Flags: []cli.Flag{
-			SolcPathFlag,
-		},
-	},
-}
-
 func init() {
 	// Override the default app help template
 	cli.AppHelpTemplate = AppHelpTemplate
