+++ dump-app-config
@@ -1,0 +1,1 @@
+Config = []
